Refactored module and updated scaffolding template sample
