--- ellar/core/conf/hidden_state_template/__ellar__/ellar/state
@@ -1,1 +1,0 @@
-{}
