--- ellar/core/conf/hidden_state_template/setup.json
@@ -1,21 +1,0 @@
-{
-  "files": [
-    {
-      "name": "__ellar__",
-      "is_directory": "true",
-      "files": [
-        {
-          "name": "ellar",
-          "is_directory": "true",
-          "files": [
-            {
-              "name": "state",
-              "is_directory": "false",
-              "files": []
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
