+++ ellar/core/conf/module_template/module_name/module.ellar
@@ -1,0 +1,37 @@
+"""
+@Module(
+    controllers=[MyController],
+    providers=[
+        YourService,
+        ProviderConfig(IService, use_class=AService),
+        ProviderConfig(IFoo, use_value=FooService()),
+    ],
+    routers=(routerA, routerB)
+    statics='statics',
+    template='template_folder',
+    # base_directory -> default is the `{{module_name}}` folder
+)
+class MyModule(ModuleBase):
+    def register_providers(self, container: Container) -> None:
+        # for more complicated provider registrations
+        pass
+
+"""
+from ellar.common import Module
+from ellar.core import ModuleBase
+from ellar.di import Container
+
+from .controllers import {{module_name | capitalize}}Controller
+
+
+@Module(
+    controllers=[{{module_name | capitalize}}Controller],
+    providers=[],
+    routers=[],
+)
+class {{module_name | capitalize}}Module(ModuleBase):
+    def register_providers(self, container: Container) -> None:
+        # for more complicated provider registrations
+        # container.register_instance(...)
+        pass
+
