+++ ellar/core/conf/module_template/module_name/routers.ellar
@@ -1,0 +1,13 @@
+"""
+Define endpoints routes in python function fashion
+example:
+
+my_router = ModuleRouter("/cats", tag="Cats", description="Cats Resource description")
+
+@my_router.get('/')
+def index(request: Request):
+    return {'detail': 'Welcome to Cats Resource'}
+"""
+
+from ellar.common import ModuleRouter
+
