+++ ellar/core/conf/module_template/module_name/schemas.ellar
@@ -1,0 +1,17 @@
+"""
+Define Serializers/DTOs
+Example:
+
+class ASampleDTO(Serializer):
+    name: str
+    age: t.Optional[int] = None
+
+for dataclasses, Inherit from DataclassSerializer
+
+@dataclass
+class ASampleDTO(DataclassSerializer):
+    name: str
+    age: t.Optional[int] = None
+"""
+from ellar.serializer import DataclassSerializer, Serializer
+
