+++ ellar/core/conf/module_template/module_name/services.ellar
@@ -1,0 +1,13 @@
+"""
+Create a provider and declare its scope
+
+@injectable
+class AProvider
+    pass
+
+@injectable(scope=transient_scope)
+class BProvider
+    pass
+"""
+from ellar.di import injectable
+
