+++ ellar/core/conf/project_template/project_name/root_module.ellar
@@ -1,10 +1,12 @@
 from ellar.common import Module, exception_handler
-from ellar.core.response import Response, JSONResponse
+from ellar.core import ModuleBase
 from ellar.core.connection import Request
+from ellar.core.response import JSONResponse, Response
 
 
 @Module()
-class {{project_name}}Module:
+class ApplicationModule(ModuleBase):
     @exception_handler(404)
     def exception_404_handler(cls, request: Request, exc: Exception) -> Response:
-        return JSONResponse(dict(detail='Resource not found.'))
+        return JSONResponse(dict(detail="Resource not found."))
+
