--- ellar/core/conf/project_template/project_name/settings.ellar
@@ -1,50 +1,0 @@
-import typing as t
-from pydantic.json import ENCODERS_BY_TYPE as encoders_by_type
-from starlette.middleware import Middleware
-from starlette.responses import JSONResponse
-from ellar.core.versioning import BaseAPIVersioning, DefaultAPIVersioning
-
-
-DEBUG: bool = True
-
-DEFAULT_JSON_CLASS: t.Type[JSONResponse] = JSONResponse
-SECRET_KEY: str = "{{secret_key}}"
-
-# injector auto_bind = True allows you to resolve types that are not registered on the container
-# For more info, read: https://injector.readthedocs.io/en/latest/index.html
-INJECTOR_AUTO_BIND = False
-
-# jinja Environment options
-# https://jinja.palletsprojects.com/en/3.0.x/api/#high-level-api
-JINJA_TEMPLATES_OPTIONS: t.Dict[str, t.Any] = {"auto_reload": DEBUG}
-
-# Application route versioning scheme
-VERSIONING_SCHEME: BaseAPIVersioning = DefaultAPIVersioning()
-
-# Enable or Disable Application Router route searching by appending backslash
-REDIRECT_SLASHES: bool = False
-
-# Define references to static folders in python packages.
-# eg STATIC_FOLDER_PACKAGES = [('boostrap4', 'statics')]
-STATIC_FOLDER_PACKAGES: t.Optional[t.List[t.Union[str, t.Tuple[str]]]] = []
-
-# Define references to static folders defined within the project
-STATIC_DIRECTORIES: t.Optional[t.List[t.Union[str, t.Any]]] = []
-
-# static route
-STATIC_MOUNT_PATH: str = "/static"
-
-# Application middlewares
-MIDDLEWARE: t.Sequence[Middleware] = []
-
-# A dictionary mapping either integer status codes,
-# or exception class types onto callables which handle the exceptions.
-# Exception handler callables should be of the form
-# `handler(request, exc) -> response` and may be be either standard functions, or async functions.
-USER_CUSTOM_EXCEPTION_HANDLERS: t.Dict[t.Union[int, t.Type[Exception]], t.Callable] = {}
-
-# Object Serializer custom encoders
-SERIALIZER_CUSTOM_ENCODER: t.Dict[t.Any, t.Callable[[t.Any], t.Any]] = encoders_by_type
-
-# ASGI Application
-ASGI_APPLICATION: t.Optional[str] = "{{project_name}}.server:application"
