+++ ellar/core/conf/project_template/tests/conftest.ellar
@@ -1,0 +1,1 @@
+from ellar.core import TestClientFactory, TestClient
